Move private functions to private files (example)
