+++ src/CMakeLists.txt
@@ -29,6 +29,7 @@ INCLUDE_DIRECTORIES(
 	"${CMAKE_BINARY_DIR}/include"
 	"${CMAKE_SOURCE_DIR}"
 	"${CMAKE_SOURCE_DIR}/include"
+	"${CMAKE_SOURCE_DIR}/src/include"
 )
 
 IF(WIN32)
